+++ jnosql-lite/mapping-lite-column-test/src/test/java/org/eclipse/jnosql/lite/mapping/entities/ColumnEntityConverterTest.java
@@ -36,7 +36,7 @@ import org.eclipse.jnosql.lite.mapping.entities.UserScope;
 import org.eclipse.jnosql.lite.mapping.entities.Vendor;
 import org.eclipse.jnosql.lite.mapping.entities.Worker;
 import org.eclipse.jnosql.lite.mapping.entities.ZipCode;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.column.ColumnEntityConverter;
 import org.eclipse.jnosql.mapping.column.spi.ColumnExtension;
 import org.eclipse.jnosql.mapping.spi.EntityMetadataExtension;
