+++ jnosql-lite/mapping-lite-processor/src/main/resources/repository_column.mustache
@@ -19,7 +19,7 @@ import jakarta.data.repository.Pageable;
 import org.eclipse.jnosql.communication.column.ColumnQuery;
 
 import org.eclipse.jnosql.lite.mapping.metadata.LiteEntitiesMetadata;
-import org.eclipse.jnosql.mapping.NoSQLPage;
+import org.eclipse.jnosql.mapping.core.NoSQLPage;
 import org.eclipse.jnosql.mapping.column.JNoSQLColumnTemplate;
 import org.eclipse.jnosql.mapping.column.MappingColumnQuery;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
