+++ jnosql-lite/mapping-lite-processor/src/main/resources/repository_document.mustache
@@ -18,7 +18,7 @@ import jakarta.data.repository.Page;
 import jakarta.data.repository.Pageable;
 import org.eclipse.jnosql.communication.document.DocumentQuery;
 import org.eclipse.jnosql.lite.mapping.metadata.LiteEntitiesMetadata;
-import org.eclipse.jnosql.mapping.NoSQLPage;
+import org.eclipse.jnosql.mapping.core.NoSQLPage;
 import org.eclipse.jnosql.mapping.document.JNoSQLDocumentTemplate;
 import org.eclipse.jnosql.mapping.document.MappingDocumentQuery;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
