+++ jnosql-lite/mapping-lite-processor/src/main/resources/repository_graph.mustache
@@ -17,7 +17,7 @@ package {{package}};
 import jakarta.data.repository.Page;
 import jakarta.data.repository.Pageable;
 import org.eclipse.jnosql.lite.mapping.metadata.LiteEntitiesMetadata;
-import org.eclipse.jnosql.mapping.NoSQLPage;
+import org.eclipse.jnosql.mapping.core.NoSQLPage;
 import org.eclipse.jnosql.mapping.graph.GraphTemplate;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.mapping.metadata.EntityMetadata;
