+++ jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/ColumnTemplateValidationTest.java
@@ -18,7 +18,7 @@ import jakarta.inject.Inject;
 import jakarta.nosql.column.ColumnTemplate;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.column.ColumnEntityConverter;
 import org.eclipse.jnosql.mapping.column.spi.ColumnExtension;
 import org.eclipse.jnosql.mapping.reflection.Reflections;
