+++ jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/ConstructorValidationTest.java
@@ -18,7 +18,7 @@ import jakarta.data.exceptions.MappingException;
 import jakarta.inject.Inject;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.reflection.ConstructorEvent;
 import org.eclipse.jnosql.mapping.reflection.Reflections;
 import org.eclipse.jnosql.mapping.spi.EntityMetadataExtension;
