+++ jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/DocumentTemplateValidationTest.java
@@ -18,7 +18,7 @@ import jakarta.inject.Inject;
 import jakarta.nosql.document.DocumentTemplate;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.document.DocumentEntityConverter;
 import org.eclipse.jnosql.mapping.document.spi.DocumentExtension;
 import org.eclipse.jnosql.mapping.reflection.Reflections;
