+++ jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/GraphTemplateValidationTest.java
@@ -17,7 +17,7 @@ package org.eclipse.jnosql.mapping.validation;
 import jakarta.inject.Inject;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.graph.GraphTemplate;
 import org.eclipse.jnosql.mapping.graph.spi.GraphExtension;
 import org.eclipse.jnosql.mapping.reflection.Reflections;
