+++ jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/KeyValueTemplateValidationTest.java
@@ -18,7 +18,7 @@ import jakarta.inject.Inject;
 import jakarta.nosql.keyvalue.KeyValueTemplate;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
-import org.eclipse.jnosql.mapping.Converters;
+import org.eclipse.jnosql.mapping.core.Converters;
 import org.eclipse.jnosql.mapping.keyvalue.KeyValueEntityConverter;
 import org.eclipse.jnosql.mapping.keyvalue.spi.KeyValueExtension;
 import org.eclipse.jnosql.mapping.reflection.Reflections;
